using float as return type

+++ consumers/token-price-service/app/schemas/token_price.py
@@ -8,7 +8,7 @@ getcontext().prec = 18
 
 class TokenPrice(BaseModel):
     resource_address: str
-    usd_price: Decimal
+    usd_price: float
     last_updated_at: datetime
 
     class Config:
